+++ src/calculator/nodes/NumberValue.java
@@ -3,22 +3,22 @@ package calculator.nodes;
 import calculator.Context;
 
 public class NumberValue extends Node {
-    private int value;
+    private double value;
 
-    public NumberValue(Context context, int value){
+    public NumberValue(Context context, double value){
         super(context);
         this.value = value;
     }
 
-    public NumberValue(int value) {
+    public NumberValue(double value) {
         this(null, value);
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
